requirements.txt
discord.py
pytz
aiohttp
buildabot